Feat(InteractionDialog): disable button by form's validation
